refactor(sheet): sequence node migration
