+++ packages/sheets-formula/src/controllers/prompt.controller.ts
@@ -53,6 +53,7 @@ import {
     deserializeRangeWithSheet,
     generateStringWithSequence,
     getAbsoluteRefTypeWitString,
+    Lexer,
     LexerTreeBuilder,
     matchRefDrawToken,
     matchToken,
@@ -159,6 +160,7 @@ export class PromptController extends Disposable {
         @Inject(IEditorBridgeService) private readonly _editorBridgeService: EditorBridgeService,
         @Inject(IFormulaPromptService) private readonly _formulaPromptService: IFormulaPromptService,
         @Inject(LexerTreeBuilder) private readonly _lexerTreeBuilder: LexerTreeBuilder,
+        @Inject(Lexer) private readonly _lexer: Lexer,
         @IRenderManagerService private readonly _renderManagerService: IRenderManagerService,
         @Inject(ThemeService) private readonly _themeService: ThemeService,
         @Inject(SelectionManagerService) private readonly _selectionManagerService: SelectionManagerService,
@@ -812,7 +814,7 @@ export class PromptController extends Disposable {
             this._contextService.setContextValue(FOCUSING_EDITOR_INPUT_FORMULA, true);
 
             const lastSequenceNodes =
-                this._lexerTreeBuilder.sequenceNodesBuilder(config.dataStream.replace(/\r/g, '').replace(/\n/g, '')) ||
+                this._lexer.sequenceNodesBuilder(config.dataStream.replace(/\r/g, '').replace(/\n/g, '')) ||
                 [];
 
             this._formulaPromptService.setSequenceNodes(lastSequenceNodes);
@@ -1908,7 +1910,7 @@ export class PromptController extends Disposable {
                         });
                     }
 
-                    const lastSequenceNodes = this._lexerTreeBuilder.sequenceNodesBuilder(formulaString) || [];
+                    const lastSequenceNodes = this._lexer.sequenceNodesBuilder(formulaString) || [];
 
                     this._formulaPromptService.setSequenceNodes(lastSequenceNodes);
