+++ packages/sheets-formula/src/services/formula-ref-range.service.ts
@@ -16,7 +16,7 @@
 
 import type { IMutationInfo, IRange } from '@univerjs/core';
 import { Disposable, DisposableCollection, IUniverInstanceService } from '@univerjs/core';
-import { deserializeRangeWithSheet, ErrorType, generateStringWithSequence, LexerTreeBuilder, sequenceNodeType, serializeRange, serializeRangeWithSheet, serializeRangeWithSpreadsheet } from '@univerjs/engine-formula';
+import { deserializeRangeWithSheet, ErrorType, generateStringWithSequence, Lexer, LexerTreeBuilder, sequenceNodeType, serializeRange, serializeRangeWithSheet, serializeRangeWithSpreadsheet } from '@univerjs/engine-formula';
 import type { EffectRefRangeParams } from '@univerjs/sheets';
 import { handleDefaultRangeChangeWithEffectRefCommands, RefRangeService } from '@univerjs/sheets';
 import type { IDisposable } from '@wendellhu/redi';
@@ -32,7 +32,7 @@ export type FormulaChangeCallback = (formulaString: string) => {
 export class FormulaRefRangeService extends Disposable {
     constructor(
         @Inject(RefRangeService) private readonly _refRangeService: RefRangeService,
-        @Inject(LexerTreeBuilder) private readonly _lexerTreeBuilder: LexerTreeBuilder,
+        @Inject(Lexer) private readonly _lexer: Lexer,
         @IUniverInstanceService private readonly _univerInstanceService: IUniverInstanceService
     ) {
         super();
@@ -40,7 +40,7 @@ export class FormulaRefRangeService extends Disposable {
 
     registerFormula(formula: string, callback: FormulaChangeCallback): IDisposable {
         const rangeMap = new Map<string, { unitId: string; subUnitId: string; range: IRange; sheetName: string }>();
-        const sequenceNodes = this._lexerTreeBuilder.sequenceNodesBuilder(formula);
+        const sequenceNodes = this._lexer.sequenceNodesBuilder(formula);
         const disposableCollection = new DisposableCollection();
         const handleChange = (params: EffectRefRangeParams) => {
             const workbook = this._univerInstanceService.getCurrentUniverSheetInstance();
